microservices/strategy-service/requirements.txt
fastapi[standard]
datetime
dotenv
pyyaml
APScheduler
pytz
backtrader
pandas
numpy
yfinance
scipy
requests
websocket-client